Add independent solution file

+++ landstalker/src/3d_maps/MapToTmx.cpp
@@ -94,7 +94,7 @@ bool MapToTmx::ImportFromTmx(const std::string& fname, Tilemap3D& map)
 		fg.size() == static_cast<std::size_t>(width * height) &&
 		bg.size() == static_cast<std::size_t>(width * height))
 	{
-		map.Resize(width, height);
+		map.Resize(static_cast<uint8_t>(width), static_cast<uint8_t>(height));
 		int i = 0;
 		for (int y = 0; y < height; ++y)
 		{
